+++ packages/contrast-colors/package.json
@@ -25,6 +25,9 @@
     "d3-hsv": "^0.1.0"
   },
   "devDependencies": {
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.23.4",
     "jest": "^26.6.3"
   },
   "publishConfig": {
